Refactor and split onto a new file
