+++ src/commands/hardis/project/deploy/smart.ts
@@ -26,7 +26,8 @@ import {
   uxLog,
 } from '../../../../common/utils/index.js';
 import { CONSTANTS, getConfig } from '../../../../config/index.js';
-import { smartDeploy, removePackageXmlContent, createEmptyPackageXml, extendPackageFileWithDependencies } from '../../../../common/utils/deployUtils.js';
+import { smartDeploy, removePackageXmlContent, createEmptyPackageXml } from '../../../../common/utils/deployUtils.js';
+import { extendPackageFileWithDependencies } from '../../../../common/utils/deltaUtils.js';
 import { isProductionOrg, promptOrgUsernameDefault, setConnectionVariables } from '../../../../common/utils/orgUtils.js';
 import { getApexTestClasses } from '../../../../common/utils/classUtils.js';
 import { listMajorOrgs, restoreListViewMine } from '../../../../common/utils/orgConfigUtils.js';
@@ -485,7 +486,7 @@ If testlevel=RunRepositoryTests, can contain a regular expression to keep only c
 
       if (process.env.EXTEND_DELTA_DEPLOYMENT === 'true' || this.configInfo.extendDeltaDeployment === true) {
         uxLog("action", this, c.cyan('[DeltaDeployment] Extending package.xml with dependencies ...'));
-        await extendPackageFileWithDependencies(diffPackageXml);
+        await extendPackageFileWithDependencies(diffPackageXml, this.packageXmlFile);
       }
 
       await removePackageXmlContent(this.packageXmlFile, diffPackageXml, true, {
